requirements.txt
pygame
numpy
sounddevice
moviepy
google-cloud-speech
openai
pydub
